@@ -1,28 +1,15 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
-import {createStore } from 'redux';
-import allReducers from './reducers';
-import { Provider } from 'react-redux';
-
-const store = createStore(
-  allReducers, 
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-);
+import reportWebVitals from './reportWebVitals';;
 
 
-ReactDOM.render(
-  <React.StrictMode>
-    
-    <Provider store={store}>
-    <App />
-    </Provider>
+// After
+const container = document.getElementById('app');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<App/>);
 
-  </React.StrictMode>,
-  document.getElementById('root')
-);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
