@@ -1,95 +1,13 @@
-import React, { Fragment, useEffect, useState } from 'react';
-/* import Duck from './Duck'; */
-import { useSelector, useDispatch } from 'react-redux';
-import { increment, decrement, login } from './actions';
-
+import React from 'react';
+import Duck from './Duck'; 
 
 
 
 const App = () => {
   return (
-    <Router>
-      <Switch>
-        <Route path="/duck">
-          <Duck />
-        </Route>
-        <Route path="/">
-          <Index />
-        </Route>
-      </Switch>
-    </Router>
+    <Duck />       
   )
 }
-
-const Index = () => {
-  useEffect(() => {
-    const getAPI = async () => {
-      const response = await fetch('http://localhost:8080/');
-      const data = await response.json();
-
-      try {
-        console.log(data);
-        setLoading(false);
-        setDuck(data);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-    getAPI();
-  }, []);
-
-  const [duck, setDuck] = useState([]);
-  const [loading, setLoading] = useState(true);
-
-  const counter = useSelector(state => state.counter);
-  const isLogged = useSelector(state => state.isLogged);
-  const dispatch = useDispatch();
-
-
-  return (
-    <Fragment>
-
-      <h1>Duck Home</h1>
-
-        <div>
-          {loading ? (
-            <div>Loading</div>
-          ) : (
-            
-            <div>
-              
-              <h1>Counter {counter}</h1>
-              <button onClick={() => dispatch(increment())}>+</button>
-              <button onClick={() => dispatch(decrement())}>-</button>
-              <hr />   
-              
-              {isLogged ? 
-                <div><button onClick={() => dispatch(login())}>Logoff</button>
-                  <h3>Valuable Information I should not see</h3></div> :
-              <button onClick={() => dispatch(login())}>Login</button>}
-              <hr />
-              
-              {duck.map((data) => (
-                <div key={data._id}>
-                  <ul>
-                    <li>
-                      <h1>
-                        {/*   <Link to={data._id}>{data.user}</Link>    */}
-                        <a href={data._id}>{data.user}</a>
-                      </h1>
-                    </li>
-                  </ul> 
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-
-
-        
-    </Fragment>
-  );
-};
 
 
 
